move content to fmt module
