+++ sdk/typespec/typespec_client_core/CHANGELOG.md
@@ -5,8 +5,7 @@
 ### Features Added
 
 - Added `#[safe]` attribute helper for `SafeDebug` derive macro to show or hide types and members as appropriate.
-- Added module `serde` which contains various helpers.
-  - Added module `as_string` which is used to (de)serialize types in string format.
+- Added module `fmt::as_string` which is used to (de)serialize types in string format.
 
 ### Breaking Changes
